Combined logger and openobserve file into a single file
