+++ app/src/openobserve.py
@@ -1,16 +1,23 @@
 """
-OpenObserve Logging Client.
+OpenObserve Logging Utility.
 
-This module provides utilities for sending event logs to an OpenObserve instance
-via HTTP POST requests using Basic Authentication.
+This module provides helper functions for sending structured event logs
+to an OpenObserve instance. It supports automatic log enrichment with
+request metadata and user context, enabling consistent tracking of
+API operations across applications.
 
-Features:
-    - Encodes credentials in Base64 for authentication.
-    - Constructs the OpenObserve endpoint URL dynamically from constants.
-    - Sends structured JSON logs to the OpenObserve stream.
+The module includes:
+    - Preparation of authentication headers for OpenObserve.
+    - Utility to send event data (`_post_log_event`) via HTTP POST.
+    - High-level function (`log_event`) that combines user and request data.
+    - Optional wrapper (`log_executive_event`) for executive-specific logs.
+
+These functions are primarily used for auditing, analytics,
+and monitoring of API activities across different application contexts.
 """
 
 import base64, json, requests
+from typing import Union
 from requests import Response
 
 from app.src.constants import (
@@ -22,6 +29,9 @@ from app.src.constants import (
     OPENOBSERVE_STREAM,
     OPENOBSERVE_USERNAME,
 )
+from app.src.db import ExecutiveToken
+from app.src.schemas import RequestInfo
+from app.src.enums import AppID
 
 # Prepare Basic Auth credentials
 credentials = base64.b64encode(
@@ -36,7 +46,7 @@ openobserve_host = f"{OPENOBSERVE_PROTOCOL}://{OPENOBSERVE_HOST}:{OPENOBSERVE_PO
 openobserve_url = f"{openobserve_host}/api/{OPENOBSERVE_ORG}/{OPENOBSERVE_STREAM}/_json"
 
 
-def post_log_event(event_data: dict) -> Response:
+def _post_log_event(event_data: dict) -> Response:
     """
     Send an event log to the configured OpenObserve instance.
 
@@ -57,7 +67,59 @@ def post_log_event(event_data: dict) -> Response:
         requests.Response: The HTTP response object returned by the OpenObserve API.
     """
     try:
-        response = requests.post(openobserve_url, headers=headers, data=json.dumps(event_data))
+        response = requests.post(
+            openobserve_url, headers=headers, data=json.dumps(event_data)
+        )
         response.raise_for_status()
     except Exception:
         pass
+
+
+def log_event(
+    token: Union[ExecutiveToken],
+    request_info: RequestInfo,
+    data: dict,
+) -> None:
+    """
+    Log an event to OpenObserve with request and user context.
+
+    Args:
+        token (Union[ExecutiveToken]): Authenticated user token.
+        request_info (RequestInfo): Metadata about the current request.
+        data (dict): Additional event-specific details to include in the log.
+
+    Notes:
+        - Automatically attaches `_app_id`, `_method`, `_path`, and user-specific ID.
+        - User-specific key depends on the app:
+            - Executive → `_executive_id`
+            - Operator  → `_operator_id`
+            - Vendor    → `_vendor_id`
+    """
+    log_details = {
+        "_method": request_info.method,
+        "_path": request_info.path,
+        "_app_id": request_info.app_id,
+    }
+
+    if request_info.app_id == AppID.EXECUTIVE and isinstance(token, ExecutiveToken):
+        log_details["_executive_id"] = token.executive_id
+
+    log_details.update(data)
+    _post_log_event(log_details)
+
+
+# Convenience wrappers (optional — use only if you want explicit naming in routes)
+def log_executive_event(
+    token: ExecutiveToken, request_info: RequestInfo, data: dict
+) -> None:
+    """
+    Convenience wrapper to log events specifically for Executive context.
+
+    Args:
+        token (ExecutiveToken): Authenticated executive token.
+        request_info (RequestInfo): Request metadata (method, path, app ID).
+        data (dict): Additional event details.
+    Returns:
+        None
+    """
+    log_event(token, request_info, data)
